+++ prepareimport/adminobject.go
@@ -13,6 +13,13 @@ type IDProperty struct {
 	Type string   `json:"type"`
 }
 
+// ParamProperty represents the "param" property of an Admin object.
+type ParamProperty struct {
+	DateDebut       MongoDate `json:"date_debut"`
+	DateFin         MongoDate `json:"date_fin"`
+	DateFinEffectif MongoDate `json:"date_fin_effectif"`
+}
+
 // UnsupportedFilesError is an Error object that lists files that were not supported.
 type UnsupportedFilesError struct {
 	UnsupportedFiles []string
